fix: 회사 삭제시 회원 탈퇴 버그 수정

+++ src/main/java/team/mjk/agent/domain/company/application/CompanyService.java
@@ -173,8 +173,12 @@ public class CompanyService {
 
     receiptService.getReceiptsByCompany(company)
         .forEach(receipt -> receiptService.deleteReceipt(memberId, receipt.getId()));
-    memberRepository.deleteAllByCompanyId(company.getId());
-    companyRepository.delete(company);
+
+    member.updateCompany(null);
+    boolean hasOtherMembers = memberRepository.existsByCompanyId(company.getId());
+    if (!hasOtherMembers) {
+      companyRepository.delete(company);
+    }
 
     return company.getId();
   }
